+++ app/components/children/formBuilder-children-old/formElement-children/InputTextElement.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+class InputTextElement extends React.Component {
+    render() {
+        return (
+            <input name={this.props.props.name} disabled type="text" />
+        )
+    }
+}
+
+export default InputTextElement
